// frontend/src/app/model/Quittance.ts
export class Quittance {
      quit_id: number;
      quit_numero: number;
      quit_Facture: number;
      quit_numeropolice: number;
      quit_numavenant: number;
      quit_numerorisque: number;
      quit_numerocie: String;
      quit_numerointermedaire: number;
      quit_typequittance: String;
      quit_typeecriture: number;
      quit_typemvt: number;
      quit_dateemission: Date;
      quit_datecomotable: Date;
      quit_dateeffet: Date;
      quit_dateecheance: Date;
      quit_typologieannulation: number;
      quit_dateannulation: Date;
      quit_numeroaperitrice: String;
      quit_primenette: number;
      quit_primeext: number;
      quit_commissionsapporteur1: number;
      quit_commissionsapporteur2: number;
      quit_accessoirecompagnie: number;
      quit_accessoireapporteur: number;
      quit_tauxte: number;
      quit_codetaxete: number;
      quit_tauxtva: number;
      quit_codetva: number;
      quit_tauxtps: number;
      quit_codetps: number;
      quit_mtntaxete: number;
      quit_mtntva: number;
      quit_mtntps: number;
      quit_mntreductionprime: number;
      quit_primettc: number;
      quit_mntprimencaisse: number;
      quit_dateencaissament: Date;
      quit_tauxreductioncommer: number;
      quit_tauxbonus: number;
      quit_tauxreductionautres: number;
      quit_mtnreduction: number;
      quit_mtnbonus: number;
      quit_mtnreductionautres: number;
      quit_numeroattestationauto: String;
      quit_numeroressp: String;
      quit_numerocertif: String;
      quit_exoneration: String;
      quit_dateexoneration: Date;
      quit_codeutilisateur: String;
      quit_numeroquittanceannul: number;
      quit_datemiseajour: Date;
      quit_datecomptabilisation: Date;
      quit_anciennumerofacture: String;
      quit_status: String;
      active: number;


}